update ui to work with multi process
Modified usage of the agent-client to work with its new version.
Updated homepage and moved all the processes' info to the process
page.
The process list has been added in the side-navigation component.
A dropdown is triggered when selecting a process to access directly
its maps and segments.
Breadcrumbs have been updated